# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrewDesk Café Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (optionally filter by session)
      parameters:
        - name: sessionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (items added to active session)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (kitchen workflow)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete (remove) an order entirely, logging it for the cancellations report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /orders/{orderId}/items/{itemId}:
    delete:
      operationId: cancelOrderItem
      tags: [orders]
      summary: Cancel a single line item in an order; logged for the cancellations report
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Item cancelled
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                  orderDeleted:
                    type: boolean
                  order:
                    $ref: "#/components/schemas/Order"

  # ── Customers ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [id, menuItemId, menuItemName, quantity, unitPrice, subtotal]
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        notes:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, sessionId, status, totalAmount, createdAt, items]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, preparing, ready, delivered, cancelled]
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        notes:
          type: string
    OrderInput:
      type: object
      required: [sessionId, items]
      properties:
        sessionId:
          type: integer
        staffId:
          type: integer
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, ready, delivered, cancelled]
        notes:
          type: string
        reason:
          type: string

    # ── Invoices ────────────────────────────────────────────────────────
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    # ── Cancellations report (manager only) ─────────────────────────────
...